move bot commands to own files
